requirements.txt
python-telegram-bot[job-queue]==21.6
anthropic==0.28.0
